refactor: 상수 파일(Constants.js)을 utils 디렉터리로 이동
