+++ src/userInterface/InputView.js
@@ -1,5 +1,5 @@
 import { Console } from '@woowacourse/mission-utils';
-import { INPUT_VIEW } from '../Constants.js';
+import { INPUT_VIEW } from '../utils/Constants.js';
 
 const inputView = {
   askPayment: async () => {
